backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-community==0.0.12
langchain-huggingface==0.0.1
python-dotenv==1.0.0
faiss-cpu==1.7.4
PyPDF2==3.0.1
requests==2.31.0
tavily-python==0.3.3
numpy==1.24.3
scikit-learn==1.3.0
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
huggingface-hub==0.19.4
ollama==0.1.7